// NextJS/src/app/types/types.ts
export type AcademyLevelEmployeeData = {
  nivel_Academico_id: number;
  nivelAcademico: {
    id: number;
    nivelacademico: string;
  } | null;
  institucion: string;
  capacitacion: string;
  carrera: {
    id: number;
    nombre_carrera: string;
  } | null;
  mension: {
    id: number;
    nombre_mencion: string;
    carrera: {
      id: number;
      nombre_carrera: string;
    };
  } | null;
};
export interface Municipality {
  id: number;
  municipio: string;
  estadoid: number;
}

export type Cargo = {
  id: number;
  cargo: string;
};

export type Nomina = {
  id: number;
  nomina: string;
  requiere_codig: boolean;
  es_activo: boolean;
};

export type OrganismosAds = {
  id: number;
  Organismoadscrito: string;
};

export type Grado = {
  id: number;
  grado: string;
};

export type Status = {
  id: number;
  estatus: string;
};
export interface States {
  id: number;
  estado: string;
  Region: {
    id: number;
    region: string;
  };
}
// ...
export interface ShirtSize {
  id: number;
  talla: string;
}
export interface PantsSize {
  id: number;
  talla: string;
}
export interface ShoesSize {
  id: number;
  talla: number;
}
export interface MaritalStatusType {
  id: number;
  estadoCivil: string;
}
export interface ParentType {
  id: number;
  descripcion_parentesco: string;
}

export interface DirectionGeneral {
  id: number;
  Codigo: string;
  direccion_general: string;
}

export interface DirectionLine {
  id: number;
  Codigo: string;
  direccion_linea: string;
}

export interface Coordination {
  id: number;
  Codigo: string;
  coordinacion: string;
}
export interface Dependency {
  id: number;
  Codigo: string;
  dependencia: string;
}

export interface DireccionGeneral {
  id: number;
  direccion_general: string;
  dependencia: {
    id: number;
    Codigo: string;
    dependencia: string;
  };
}
export interface DireccionLinea {
  id: number;
  Codigo: string;
  direccion_linea: string;
}

export interface Municipality {
  id: number;
  municipio: string;
}
export interface DewllingInfo {
  estado: States;
  municipio: Municipality;
  parroquia: Parish;
  direccion_exacta: string;
  condicion: {
    id: number;
    condicion: string;
  };
}
export interface HealthProfile {
  grupoSanguineo: {
    id: number;
    grupoSanguineo: string;
  } | null;
  alergias:
    | {
        id: number;
        alergia: string;
        categoria: {
          id: number;
          nombre_categoria: string;
        };
      }[]
    | null;
  discapacidad:
    | {
        id: number;
        discapacidad: string;
        categoria: {
          id: number;
          nombre_categoria: string;
        };
      }[]
    | null;
  patologiasCronicas:
    | {
        id: number;
        patologia: string;
        categoria: {
          id: number;
          nombre_categoria: string;
        };
      }[]
    | null;
}

export interface Background {
  id: number;
  institucion: string | null;
  fecha_ingreso: string | null;
  fecha_egreso: string | null;
  fecha_actualizacion: string | null;
}
export interface PhysicalProfile {
  tallaCamisa: ShirtSize | null;
  tallaPantalon: PantsSize | null;
  tallaZapatos: ShoesSize | null;
}
export interface InfoCode {
  id: number;
  codigo: string;
  denominacioncargo: Cargo;
  denominacioncargoespecifico: Cargo;
  grado: Grado;
  tiponomina: Nomina;
  OrganismoAdscrito: OrganismosAds | null;
  Dependencia: Dependency;
  DireccionGeneral: DireccionGeneral | null;
  DireccionLinea: DireccionLinea | null;
  Coordinacion: Coordination | null;
  estatusid: Status;
  observaciones: string | null;
  fecha_actualizacion: string;
}
// ...
export interface EmployeeData {
  anos_apn: number;
  id: number;
  cedulaidentidad: string;
  nombres: string;
  apellidos: string;
  profile: string;
  fecha_nacimiento: string;
  n_contrato: string;
  sexo: Sex;
  estadoCivil: StatusCivil;
  datos_vivienda: DewllingInfo;
  perfil_salud: HealthProfile;
  perfil_fisico: PhysicalProfile;
  formacion_academica: AcademyLevelEmployeeData;
  antecedentes: Background[];
  fechaingresoorganismo: string;
  fecha_actualizacion: string;
  asignaciones: InfoCode[];
  correo: string | null;
  telefono_habitacion: string | null;
  telefono_movil: string | null;
}
export interface Sex {
  id: number;
  sexo: string;
}
export interface StatusCivil {
  id: number;
  estadoCivil: string;
}
// ...
export interface Parish {
  id: number;
  parroquia: string;
}